Updated material shader to use texture array and sampler. Fixed material rendering code, allowing use of diffuse textures.

+++ VulkanApp/mesh.h
@@ -20,7 +20,7 @@ struct Vertex
 	glm::vec3 pos;
 	glm::vec2 tex_coord;
 	glm::vec3 normal;
-	float mat_index;
+	uint32_t mat_index;
 
 	static VkVertexInputBindingDescription GetBindingDescription()
 	{
